+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix incorrect message suggestion to check upload status if non-default profile is used
+
 ## [1.3.2] - 2023-07-05
 
 ### Fixed